+++ packages/atomic/cypress/integration/facets/numeric-facet/numeric-facet-assertions.ts
@@ -1,4 +1,3 @@
-import {TestFixture} from '../../../fixtures/test-fixture';
 import {should} from '../../common-assertions';
 import {NumericFacetSelectors} from './numeric-facet-selectors';