+++ packages/atomic/cypress/integration/facets/rating-facet/rating-facet-assertions.ts
@@ -1,4 +1,3 @@
-import {TestFixture} from '../../../fixtures/test-fixture';
 import {
   BaseFacetSelector,
   FacetWithLinkSelector,