+++ packages/atomic/src/components.d.ts
@@ -335,6 +335,7 @@ export namespace Components {
     interface AtomicQuerySummary {
         /**
           * Whether to display the duration of the last query execution.
+          * @deprecated Use the `duration` part.
          */
         "enableDuration": boolean;
     }
@@ -1583,6 +1584,7 @@ declare namespace LocalJSX {
     interface AtomicQuerySummary {
         /**
           * Whether to display the duration of the last query execution.
+          * @deprecated Use the `duration` part.
          */
         "enableDuration"?: boolean;
     }