+++ packages/headless/src/features/result-preview/result-preview-actions.ts
@@ -1,23 +1,29 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import {HtmlRequestOptions} from '../../api/search/html/html-request';
-import {
-  AsyncThunkSearchOptions,
-  isErrorResponse,
-} from '../../api/search/search-api-client';
+import {isErrorResponse} from '../../api/search/search-api-client';
 import {
   buildResultPreviewRequest,
   StateNeededByHtmlEndpoint,
 } from './result-preview-request-builder';
+import {ClientThunkExtraArguments} from '../../app/thunk-extra-arguments';
+import {HtmlApiClient} from '../../api/search/html/html-api-client';
+import {AsyncThunkOptions} from '../../app/async-thunk-options';
+import {SearchAPIErrorWithStatusCode} from '../../api/search/search-api-error-response';
 
 interface FetchResultContentResponse {
   content: string;
   uniqueId: string;
 }
 
+export interface AsyncThunkGlobalOptions<T>
+  extends AsyncThunkOptions<T, ClientThunkExtraArguments<HtmlApiClient>> {
+  rejectValue: SearchAPIErrorWithStatusCode;
+}
+
 export const fetchResultContent = createAsyncThunk<
   FetchResultContentResponse,
   HtmlRequestOptions,
-  AsyncThunkSearchOptions<StateNeededByHtmlEndpoint>
+  AsyncThunkGlobalOptions<StateNeededByHtmlEndpoint>
 >(
   'resultPreview/fetchResultContent',
   async (options: HtmlRequestOptions, {extra, getState, rejectWithValue}) => {